move alloc-to-global pass out of frontend
squash alloc
